parser: support parsing map type (to match clickhouse)
Right now the AST that is generated is JSONExtract, which makes
some sense, as really it is extracting one key out of a JSON blob
that happens to be a map. If we decide to implement the map type
more generally then we will need to most likely add our own custom
mysql type to the parse tree.